Move GetPrefilledChannelUseCase to domain module
