+++ app/src/main/java/dev/punitd/feedly/onboarding/OnboardingViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.punitd.data.Channel
+import dev.punitd.domain.usecase.GetPrefilledChannelsUseCase
 import dev.punitd.domain.usecase.GetSavedChannelsUseCase
 import dev.punitd.domain.usecase.SaveChannelUseCase
-import dev.punitd.feedly.onboarding.domain.usecase.GetPrefilledChannelsUseCase
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
