--- app/src/main/java/dev/punitd/feedly/onboarding/domain/usecase/GetPrefilledChannelsUseCase.kt
@@ -1,7 +1,0 @@
-package dev.punitd.feedly.onboarding.domain.usecase
-
-import dev.punitd.data.Channel
-
-interface GetPrefilledChannelsUseCase {
-    suspend fun execute(): List<Channel>
-}
